requirements.txt
bcrypt==3.1.3
cffi==1.10.0
cloudinary==1.8.0
dj-database-url==0.4.2
Django==1.11.4
django-debug-toolbar==1.8
django-decorator-include==1.3
django-dotenv==1.4.1
django-filter==1.0.4
django-redis==4.8.0
djangorestframework==3.6.3
gunicorn==19.7.1
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
mock==2.0.0
pbr==3.1.1
ply==3.4
psycopg2==2.7.3
pycparser==2.18
pytz==2017.2
redis==2.10.6
six==1.10.0
slimit==0.8.1
sqlparse==0.2.3
urllib3==1.22
WebOb==1.7.3
whitenoise==3.3.0
wsgi-basic-auth==1.1.0
